+++ packages/vue-components/src/index.js
@@ -4,6 +4,7 @@ import 'floating-vue/dist/style.css';
 /* eslint-enable import/no-extraneous-dependencies */
 
 // Custom / modified components and components from yuche/vue-strap
+import box from './Box.vue';
 import closeable from './directives/Closeable';
 import dropdown from './Dropdown.vue';
 import navbar from './Navbar.vue';
@@ -20,7 +21,6 @@ import tab from './Tab.vue';
 import tabGroup from './TabGroup.vue';
 import tabset from './Tabset.vue';
 import thumbnail from './Thumbnail.vue';
-import tipBox from './TipBox.vue';
 import trigger from './Trigger.vue';
 import siteNav from './SiteNav.vue';
 import submenu from './Submenu.vue';
@@ -33,7 +33,7 @@ import tooltip from './Tooltip.vue';
 import modal from './Modal.vue';
 
 const components = {
-  box: tipBox,
+  box,
   dropdown,
   navbar,
   panel,
